requirements.txt
pandas==2.0.0
scikit
numpy==1.24.2
matplotlib==3.7.1  # 3.6.3
paho-mqtt~=1.6.1
Pillow==9.5.0
opencv-python==4.7.0.72
gymnasium==0.28.1
networkx~=3.0
seaborn~=0.12.2
coloredlogs~=15.0.1
scipy==1.10.1
scikit-learn~=1.2.2
pip~=23.0.1
attrs~=22.2.0
wheel~=0.38.4
pytz~=2022.7.1
pytest~=7.2.1
hypothesis~=6.68.1
setuptools~=60.2.0
mpmath~=1.2.1
sympy~=1.11.1
python-dateutil~=2.8.2
cycler~=0.11.0
six~=1.16.0
contourpy~=1.0.7
fonttools~=4.38.0
pygame~=2.3.0
pyparsing~=3.0.9
packaging~=23.0
kiwisolver~=1.4.4
cloudpickle~=2.2.1
humanfriendly~=10.0
pyreadline3~=3.4.1